+++ src/models/interfaces/AssetParentInterface.php
@@ -3,6 +3,7 @@
 namespace davidhirtz\yii2\media\models\interfaces;
 
 use davidhirtz\yii2\skeleton\db\ActiveQuery;
+use davidhirtz\yii2\skeleton\models\interfaces\TypeAttributeInterface;
 use yii\db\ActiveRecordInterface;
 
 /**
@@ -10,7 +11,7 @@ use yii\db\ActiveRecordInterface;
  * @property AssetInterface[] $assets {@see static::getAssets()}
  * @method string formName()
  */
-interface AssetParentInterface extends ActiveRecordInterface
+interface AssetParentInterface extends ActiveRecordInterface, TypeAttributeInterface
 {
     public function getAssets(): ActiveQuery;
     public function getAssetSizes(): ?string;